Removal of the old quad tree code
