+++ lib/Collider.js
@@ -63,11 +63,6 @@ export class Collider extends ShapedAssetEntity {
         this.velocity._perfX = vectorN.x * amplitude;
         this.velocity._perfY = vectorN.y * amplitude;
 
-        //should this be here? - Andrew
-        if (this.velocity.x != this.velocity.prevX || this.velocity.y != this.velocity.prevY) {
-            this.quadValue = this.#pen.quadTree.getValue(this.x, this.y, this.radius);
-        }
-
         super.move();
 
         this.#updateBounds();
