+++ lib/Group.js
@@ -1,5 +1,5 @@
 import { Collider } from "./Collider.js";
-import { QuadTree2 } from "./QuadTree2.js";
+import { QuadTree } from "./QuadTree.js";
 
 export function makeGroup(...elements) {
     const nuGroup = new Group(...elements);
@@ -26,7 +26,7 @@ export class Group extends Array {
             this.push(element);
         }
         this.name = null;
-        this.#quadTree = new QuadTree2();
+        this.#quadTree = new QuadTree();
     }
 
     /**
