+++ lib/Pen.js
@@ -13,8 +13,6 @@ import { lock } from "./Lock.js";
 
 import { TimeManager } from "./TimeManager.js"
 
-import { QuadTree } from "../../lib/QuadTree.js";
-
 //entities
 import { Img, ImgWrapper } from "./Img.js";
 import { Button } from "./Button.js";
@@ -88,8 +86,6 @@ export class Pen {
         
         this.#paused=false;
 
-        this.quadTree = new QuadTree(512, 512);
-
         //collections
         this.#userFuncs = {
             draw: draw,
@@ -158,7 +154,6 @@ export class Pen {
      */
     set width(value) {
         this.canvas.width = value;
-        this.quadTree = new QuadTree(value, this.h);
     }
     get width() {
         return this.canvas.width;
@@ -168,7 +163,6 @@ export class Pen {
      */
     set height(value) {
         this.canvas.height = value;
-        this.quadTree = new QuadTree(this.w, value);
     }
     get height() {
         return this.canvas.height;
