+++ lib/QuadTree.js
@@ -1,9 +1,7 @@
 export class Quad {
 
     #depth;
-    #depthInverse;
-    #width;
-    #height;
+    #size;
 
     #left;
     #right;
@@ -15,29 +13,67 @@ export class Quad {
     #quad3;
     #quad4;
     
-    constructor(depth, depthMax, width, height, x, y) {
+    #objects;
+
+    constructor(depth, depthMax, objects, size, x, y) {
         
         this.#depth = depth;
-        this.#depthInverse = depthMax - this.#depth;
-        this.#width = width;
-        const halfWidth = this.#width / 2;
-        this.#height = height;
-        const halfHeight = this.#height / 2;
+        this.#size = size;
+        const halfSize = this.#size / 2;
 
         this.#left = x;
-        this.#right = x + width;
+        this.#right = x + size;
         this.#top = y;
-        this.#bottom = y + height;
-        
-        if (depth < depthMax) {
-            if (depth == 0) {
-                this.#quad1 = new Quad(depth+1, depthMax, width, halfHeight, this.#left, this.#top);
-                this.#quad2 = new Quad(depth+1, depthMax, width, halfHeight, this.#left, this.#top + halfHeight);
+        this.#bottom = y + size;
+
+        if (objects.length <= 4 || depth == depthMax) {
+            this.#objects = objects;
+        } else {
+            this.#objects = null;
+
+            let obj1 = new Array();
+            let obj2 = new Array();
+            let obj3 = new Array();
+            let obj4 = new Array();
+            for (let i = 0; i < objects.length; i++) {
+                const o = objects[i];
+                if (o.x1 < this.#left + halfSize) { //left
+                    if (o.y1 < this.#top + halfSize) { //left-top
+                        obj1.push(o);
+                    } 
+                    if (o.y2 >= this.#top + halfSize) { //right-top
+                        obj3.push(o);
+                    }
+                }
+                if (o.x2 >= this.#left + halfSize) { //right
+                    if (o.y1 < this.#top + halfSize) { //left-bottom
+                        obj2.push(o);
+                    }
+                    if (o.y2 >= this.#top + halfSize) { //right-bottom
+                        obj4.push(o);
+                    }
+                }
+            }
+
+            if (obj1.length > 0) {
+                this.#quad1 = new Quad(depth+1, depthMax, obj1, halfSize, this.#left, this.#top);
             } else {
-                this.#quad1 = new Quad(depth+1, depthMax, halfWidth, halfHeight, this.#left, this.#top);
-                this.#quad2 = new Quad(depth+1, depthMax, halfWidth, halfHeight, this.#left + halfWidth, this.#top);
-                this.#quad3 = new Quad(depth+1, depthMax, halfWidth, halfHeight, this.#left, this.#top + halfHeight);
-                this.#quad4 = new Quad(depth+1, depthMax, halfWidth, halfHeight, this.#left + halfWidth, this.#top + halfHeight);
+                this.#quad1 = null;
+            }
+            if (obj2.length > 0) {
+                this.#quad2 = new Quad(depth+1, depthMax, obj2, halfSize, this.#left + halfSize, this.#top);
+            } else {
+                this.#quad2 = null;
+            }
+            if (obj3.length > 0) {
+                this.#quad3 = new Quad(depth+1, depthMax, obj3, halfSize, this.#left, this.#top + halfSize);
+            } else {
+                this.#quad3 = null;
+            }
+            if (obj4.length > 0) {
+                this.#quad4 = new Quad(depth+1, depthMax, obj4, halfSize, this.#left + halfSize, this.#top + halfSize);
+            } else {
+                this.#quad4 = null;
             }
         }
     }
@@ -54,46 +90,6 @@ export class Quad {
         }
     }
 
-    getValue(x1, x2, y1, y2) {
-
-        if (x2 < this.#left) {
-            return 0;
-        }
-        if (x1 > this.#right) {
-            return 0;
-        }
-        if (y2 < this.#top) {
-            return 0;
-        }
-        if (y1 > this.#bottom) {
-            return 0;
-        }
-
-        if (this.#quad1 != null) {
-
-            if (this.#depthInverse == 1) {
-                const q4 = this.#quad4.getValue(x1, x2, y1, y2) << 3;
-                const q3 = this.#quad3.getValue(x1, x2, y1, y2) << 2;
-                const q2 = this.#quad2.getValue(x1, x2, y1, y2) << 1;
-                const q1 = this.#quad1.getValue(x1, x2, y1, y2) << 0;
-                return q1 + q2 + q3 + q4;
-            } else if (this.#depthInverse == 2) {
-                const q4 = this.#quad4.getValue(x1, x2, y1, y2) << 12;
-                const q3 = this.#quad3.getValue(x1, x2, y1, y2) << 8;
-                const q2 = this.#quad2.getValue(x1, x2, y1, y2) << 4;
-                const q1 = this.#quad1.getValue(x1, x2, y1, y2) << 0;
-                return q1 + q2 + q3 + q4;
-            } else if (this.#depthInverse == 3) {
-                const q2 = this.#quad2.getValue(x1, x2, y1, y2) << 16;
-                const q1 = this.#quad1.getValue(x1, x2, y1, y2) << 0;
-                return q1 + q2;
-            }
-
-        } else {
-            return 1;
-        }
-    }
-
     get left() {
         return this.#left;
     }
@@ -106,32 +102,29 @@ export class Quad {
     get bottom() {
         return this.#bottom;
     }
-    get width() {
-        return this.#width;
+    get size() {
+        return this.#size;
     }
-    get height() {
-        return this.#height;
+    get objects() {
+        return this.#objects;
     }
 }
 
 export class QuadTree {
 
     #quad;
-    
-    constructor(width, height) {
-        const depthMax = 3;
-        this.#quad = new Quad(0, depthMax, width, height, 0, 0);
+
+    constructor() {
+        this.#quad = null;
     }
 
-    getValue(x, y, radius) {
-        const x1 = Math.round(x - radius);
-        const x2 = Math.round(x + radius);
-        const y1 = Math.round(y - radius);
-        const y2 = Math.round(y + radius);
-        return this.#quad.getValue(x1, x2, y1, y2);
+    makeTree(objects) {
+        const size = 1024;
+        const depthMax = 6;
+        this.#quad = new Quad(0, depthMax, objects, size, 0, 0);
     }
 
-    getQuad() {
+    getTree() {
         return this.#quad;
     }
 }
